ENGR00223797 MX6 SabreSD:Add fb0 base addr to kernel cmdline
This patch adds 'fb0base=0x27b00000' to kernel command line
for MX6 SabreSD platform to specify fb0 base address to be
0x27b00000, so that kernel can use a reserved fb memory
starting from this address. This address is used in uboot
fb. To support smooth UI transition from uboot to kernel,
this fb memory has to be reserved by kernel, otherwise,
it is likely to be overwritten by others, so kernel will
go on using this block of memory to save memory usage.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>