MA-9392 [Android_MX7ULP-EVK] backlight: change the init.rc instead of dts file to eliminate error log of getting max_brightness file failed
In previous method of solve the backlight issue, the backlight node name of dts file is changed to "pwm-backlight".
Because in /hardware/imx/lights/fsl_lights.c file, some function will open "/max_brightness" file with the default path.
Here we change the default backlight device hw.backlight.dev to "backlight", so the path will be
/sys/class/backlight/backlight/max_brightness and it match the default device tree.

Change-Id: Ifb2c7a5c74da858e6a27dedbfa4d225b73c3849d
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1637
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>