MA-9548 [Android_6DL_SD]Storage: External SDCard can't be recognized and it impacts the CTS testing. 100%
If we boot up emmc image, we use the SD2 as external sdcard.
we use the same image for both sd&emmc.
So we can't use the same fstab for sabresd6dq emmc&sd.

Add a fstab.freescale.storage which record the external storage.
This file will be parsed by vold.

Change-Id: I4dc89c684293a703c4b6ff0d27b4037bfe7ccbb3
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1981
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>
Reviewed-by: zhang bo <bo.zhang@nxp.com>